Tests/LibWeb/Text/expected/caret-position-line-edge-ranking.txt

below monetization stays in description text: true
below monetization does not snap to following link: true
above status line snaps to logical line start: true
